requirements.txt
alembic==1.6.5
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==8.0.1
colorama==0.4.4
cryptography==3.4.7
Flask==2.0.1
Flask-Migrate==3.0.1
flask-ngrok==0.0.25
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
Mako==1.1.4
MarkupSafe==2.0.1
pycparser==2.20
PyMySQL==1.0.2
python-dateutil==2.8.1
python-dotenv==0.17.1
python-editor==1.0.4
requests==2.25.1
six==1.16.0
SQLAlchemy==1.4.17
urllib3==1.26.5
Werkzeug==2.0.1
